% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenaterFederation.R
\docType{class}
\name{RenaterFederation}
\alias{RenaterFederation}
\title{RenaterFederation}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an RenaterFederation
}
\description{
RenaterFederation

RenaterFederation
}
\note{
Main user class to be used with \pkg{renatR}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{federation}
\section{Super class}{
\code{\link[renatR:renatRLogger]{renatR::renatRLogger}} -> \code{RenaterFederation}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{federation id}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RenaterFederation-new}{\code{RenaterFederation$new()}}
\item \href{#method-RenaterFederation-getIdentityProviders}{\code{RenaterFederation$getIdentityProviders()}}
\item \href{#method-RenaterFederation-clone}{\code{RenaterFederation$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="renatR" data-topic="renatRLogger" data-id="ERROR"><a href='../../renatR/html/renatRLogger.html#method-renatRLogger-ERROR'><code>renatR::renatRLogger$ERROR()</code></a></span></li>
<li><span class="pkg-link" data-pkg="renatR" data-topic="renatRLogger" data-id="INFO"><a href='../../renatR/html/renatRLogger.html#method-renatRLogger-INFO'><code>renatR::renatRLogger$INFO()</code></a></span></li>
<li><span class="pkg-link" data-pkg="renatR" data-topic="renatRLogger" data-id="WARN"><a href='../../renatR/html/renatRLogger.html#method-renatRLogger-WARN'><code>renatR::renatRLogger$WARN()</code></a></span></li>
<li><span class="pkg-link" data-pkg="renatR" data-topic="renatRLogger" data-id="getClass"><a href='../../renatR/html/renatRLogger.html#method-renatRLogger-getClass'><code>renatR::renatRLogger$getClass()</code></a></span></li>
<li><span class="pkg-link" data-pkg="renatR" data-topic="renatRLogger" data-id="getClassName"><a href='../../renatR/html/renatRLogger.html#method-renatRLogger-getClassName'><code>renatR::renatRLogger$getClassName()</code></a></span></li>
<li><span class="pkg-link" data-pkg="renatR" data-topic="renatRLogger" data-id="logger"><a href='../../renatR/html/renatRLogger.html#method-renatRLogger-logger'><code>renatR::renatRLogger$logger()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RenaterFederation-new"></a>}}
\if{latex}{\out{\hypertarget{method-RenaterFederation-new}{}}}
\subsection{Method \code{new()}}{
Initialize the federation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RenaterFederation$new(id = c("renater", "edugain", "test"), logger = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{federation id}

\item{\code{logger}}{logger}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RenaterFederation-getIdentityProviders"></a>}}
\if{latex}{\out{\hypertarget{method-RenaterFederation-getIdentityProviders}{}}}
\subsection{Method \code{getIdentityProviders()}}{
Get identity providers (IdPs)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RenaterFederation$getIdentityProviders(pretty = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{if output should be prettified as \link{data.frame}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
an object of class \link{list} or \link{data.frame}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RenaterFederation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RenaterFederation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RenaterFederation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
