% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renatR.R
\docType{package}
\name{renatR}
\alias{renatR}
\alias{renatR-package}
\title{Tools to connect to Renater}
\description{
Provides tools to connect to the Renater federation.
}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}
