% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renatR_logger.R
\docType{class}
\name{renatRLogger}
\alias{renatRLogger}
\title{renatRLogger}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling a simple logger
}
\description{
renatRLogger

renatRLogger
}
\note{
Logger class used internally by \pkg{renatR}
}
\keyword{logger}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{verbose.info}}{logger info status}

\item{\code{verbose.debug}}{logger debug status}

\item{\code{loggerType}}{Logger type, either "INFO", "DEBUG" or NULL (if no logger)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-renatRLogger-logger}{\code{renatRLogger$logger()}}
\item \href{#method-renatRLogger-INFO}{\code{renatRLogger$INFO()}}
\item \href{#method-renatRLogger-WARN}{\code{renatRLogger$WARN()}}
\item \href{#method-renatRLogger-ERROR}{\code{renatRLogger$ERROR()}}
\item \href{#method-renatRLogger-new}{\code{renatRLogger$new()}}
\item \href{#method-renatRLogger-getClassName}{\code{renatRLogger$getClassName()}}
\item \href{#method-renatRLogger-getClass}{\code{renatRLogger$getClass()}}
\item \href{#method-renatRLogger-clone}{\code{renatRLogger$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-renatRLogger-logger"></a>}}
\if{latex}{\out{\hypertarget{method-renatRLogger-logger}{}}}
\subsection{Method \code{logger()}}{
internal logger function for the \pkg{renatR} manager
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{renatRLogger$logger(type, text)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{logger message type, "INFO", "WARN", or "ERROR"}

\item{\code{text}}{log message}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-renatRLogger-INFO"></a>}}
\if{latex}{\out{\hypertarget{method-renatRLogger-INFO}{}}}
\subsection{Method \code{INFO()}}{
internal INFO logger function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{renatRLogger$INFO(text)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{log message}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-renatRLogger-WARN"></a>}}
\if{latex}{\out{\hypertarget{method-renatRLogger-WARN}{}}}
\subsection{Method \code{WARN()}}{
internal WARN logger function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{renatRLogger$WARN(text)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{log message}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-renatRLogger-ERROR"></a>}}
\if{latex}{\out{\hypertarget{method-renatRLogger-ERROR}{}}}
\subsection{Method \code{ERROR()}}{
internal ERROR logger function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{renatRLogger$ERROR(text)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{log message}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-renatRLogger-new"></a>}}
\if{latex}{\out{\hypertarget{method-renatRLogger-new}{}}}
\subsection{Method \code{new()}}{
initialize the \pkg{renatR} logger
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{renatRLogger$new(logger = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{logger}}{logger type NULL, 'INFO', or 'DEBUG'}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-renatRLogger-getClassName"></a>}}
\if{latex}{\out{\hypertarget{method-renatRLogger-getClassName}{}}}
\subsection{Method \code{getClassName()}}{
Get object class name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{renatRLogger$getClassName()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the class name, object of class \code{character}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-renatRLogger-getClass"></a>}}
\if{latex}{\out{\hypertarget{method-renatRLogger-getClass}{}}}
\subsection{Method \code{getClass()}}{
Get object class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{renatRLogger$getClass()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the class, object of class \code{R6}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-renatRLogger-clone"></a>}}
\if{latex}{\out{\hypertarget{method-renatRLogger-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{renatRLogger$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
